third_party/blink/web_tests/printing/allowed-page-breaks-expected.txt

Test for allowed page breaks

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: page number of "page1" is 1
FAIL: expected page number of "page2" is 2. Was 1
PASS: page number of "page3-1" is 3
PASS: page number of "page3-2" is 3

PASS successfullyParsed is true

TEST COMPLETE
